cc1: Don't use strlen concatenating strings
Adjacents strings must be concatenated, but the process was assuming
strings without embeddeding NUL characters (using strlen), but this
is not true as they can be embedded using character escape sequences.
For that reason we have to use the number of elements defined in the
type associated to the string constant.