cc1: Fix sizeof expressions
The commit c863f707 tried to fix expressions like:

	sizeof((int) 1)

and changed the reduction in sizeexp() from unary to cast,
but it had the side effect of allowing things like

	sizeof (int) 1

that is invalid, and having problems with things like

	sizeof "hello" - 1

The commit 4a0c2ccd tried to fix this last case reverting c863f707
that moved the code to the original point. The problem is that
to avoid the ambiguity when a '(' follows a sizeof (because
it can be a typename or an expression) then the rule chain

	unary
		-> SIZEOF unary
		-> SIZEOF postfix
		-> SIZEOF primary
		-> SIZEOF ( expr )

was collapsed to

	unary
		-> SIZEOF ( unary )

because when a ( happend after a SIZEOF token then the next token
was looked to see if it was a typename or not to decide to apply
the chain:

	unary
		-> SIZEOF ( typename )

And it can be seen the problem is that the collapsing was wrong
because we should have SIZEOF ( expr ) instead of SIZEOF ( unary ),
but still SIZEOF unary when no ( follows the SIZEOF token.