build: Use -fno-stack-protector only for libc
We require no pie and without stack protector only for the libc,
because we want to link the libc with the output of qbe that is
not pie. Also, the libc does not support the symbols required
by the stack protector functionality from gcc, and for that reason
we have to disable it.