cc1: Fix empty param list in newtype()
When a function does not have parameters (a void function)
then tp->n.elem is 0, and it was calling malloc with a size
of 0 and then later in eqfunc() we are accessing the pointer
returned by malloc(). It happens to work in linux, but
obviously, it was wrong.

Also, the malloc() call didn't allocate space for the NULL
pointer that must finish the list and it was expected by
eqfunc().