tests/libc: Select path to gcc-scc
The Makefile was using gcc-scc expecting it to be in the PATH, but
the correct soluction is to select the tool from the bin directory.
Also, gcc-scc was using gcc, that it is not shipped anymore with
OpenBSD, and in this context we can assume that the system compiler
is cc.