tests/cc: Update twice in update.sh
When running update.sh it can add/remove lines to the header
that makes invalid all the numbers in the first execution,
forcing to run it 2 times. As there were sereval commits updating
several tests with this problem then is better to run the update
twice in the onw script.