make: Enable use of expansions in items
An item can be composed by the concatenation of digits and the
expansion of a macro. This change smells a bit, because we are
duplicating the call to expand in too many different places,
but the different behaviour of rules, actions and assignments
drives to something like this, or a nextc() function with
too many globals and conditions.