build: Remove src dependency to include/bits/scc
This dependency is managed by the top level Makefile that ensures
it is fulfil before using main.mk. This avoid problems in cases
where the time stamp does not have enough precision to differentiate
the timestamp of the file and the directory containing it. This is
the reason why usually putting time dependencies between directories
is a bad idea.