cc2: Rewrite addstmt() and delstmt()
Addstmt() was buggy and it only worked when new statements were
added at the end, that it was always the case. As we expect that
we are going to need a pointer to the last statement then we moved
aways from the u.stmt field and we represent the function with
a specific struct, that eventually will contain curstmt too.