cc2: Detect unfinished function bodies
The parser was exiting for end of file and it could happen while
we were parsing a function, accepting functions without the
closing brace. Also, several of the global variables that mark
the state of the parser were not initialized at the beginning
of every parsing, openning the door for many different problems.

Also, moreinput() could go in a infinite loop because in case of
EOF only checked for feof() but it could happen due to some error
and have an infinite loop.

And finally, we were through the full main loop every time parser()
finished, even if we didn't parse a function, having to add
continous checks everywhere because curfun could be NULL. This
didn't make sense, and a single check is done in the main loop
keeping safe the function related functions.