cc2: Change left/right nodes naming
The name rp is used now for range pointer and it collides with
rp for right pointer. As we were not consistent because we used
rp/lp sometimes for double pointers and r/l for normal pointers
and r/l is the naming used in cc1 it makes more sense to use r/l.