cc2: Remove stub implementations
The amd64 and i386 implementations are only stubs,
and due to lack of time it is likely that they are
going to remain as stubs for long time, so it is
better to remove them and have less code to maintain.
It is likely that many of the code that it is in target
directories will be refactored to be common.