cc2: First implementation of direct switch
The fastest implementation of a switch is using a jump table
although it has some initial overhead and it can be very
inneficient in space in some cases. For this reason it is used
only in some cases. This commit is a first step that dumps
a jump table but it still don't use it.