cc2: Rewrite switch data structure
It is better to use an array for the switch table because it makes
easier to sort the values that it is something required for several of
the implementations for switch using jummp tables.

Also, as this transformation is done in sethi() it can be reused for
differnt targets without duplication. Currently, as qbe does not support
computed gotos any jump table is disabled when the name of the backend
has the substring qbe.