cc2: Copy bool() to sethi()
The bool generator is going to be needed for every target, so it is
better to do this operation in sethi() that is shared between all
the backends. Also, moving these functions to sethi() would make
possible to apply liveness analysis to the tree code in the future.

A copy of bool() is still kept in qbe/cgen.c because it is used
when boolean operators are used in expressions that is something
to be expected moved to sethi() too in a near future.