cc2: Deal with comma in sethi()
The comma operator just marks an order between different
statements, and it can be resolved with a tree transformation
moving the comma subexpressions before the current statement.
It works for any combination of operands, because the evaluation
order is not speficied an din the cases where it is speficified
like for example in boolean operators they already define
synchronization points.