make/posix: Fix signal handling
The behaviour of signal() in POSIX is underspecified, and it can restart
interrupted syscalls, so the only way to be sure that syscalls will
be interrupted by signals is to use sigaction(). As we don't want to
pollute the code with these weirdness things we just to a lazy signal
initialization in the POSIX case while we keep normal signal for
everyone else.