libc: Remove strchr(), strcmp() and strcpy() from common
These functions can be provided as architectural dependent
functions, and it happened that amd64 had a c and an assembler
version.

For sure that some architectures are going to be broken after this
commit because they expect the C version, but that would raise soon
so it is not a big issue now.