dump: Rename from objdump
This tool was just a scam that had nothing in common
with objdump. We are in a moment where having a objdump
tool makes sense, so we can begin to have a proper one
and meanwhile we rename objdump to dump in order to
iterate over objdump without breaking the tests.