build: Make libmach recursive
Libmach was the last Makefile compiling things out of the
current directory, and it was a bit ugly because it had to
define files per directory and the dependencies were harder
to generate.

As we already have a dep target that creates a new makefile
that target was reused to include the list of objects. To
be sure that dep is always called a top level deps target
was created that uses a stamp file to ensure it.