libmach: Add offset and load to Section
The load address in the address where the section is going
to be loaded that can be different to the base address where
the relocations are performed. In any UNIX system both addresses
are going to have the same value, but it could be different in
other systems.

In the case of coff the physical address is used as load address
as it does not have any other meaning in UNIX systems.

While elf support also virtual and physical addresses, it only
supports it for segments, and the current implementation of
getsec for elf only deals with sections.