libmach: Add hack to know if a section has relocations
Copying back the number of relocations works only in coff,
because it is not easy to get that data in other formats
like a.out or elf. As getsec() is used in many different
contexts it does not make sense to make it more complex
and slow only to satisfy objdump. This is just a fast
way to get some output in the case of coff, but we do
know that it will have to be changed in the future.