libmach/coff32: Add write right to text sections
Coff does not have the concept of read and write rights,
and assuming that text sections would be not writable was
a too strong assumption. It is better to assume that they
can be modified and some systmes will disallow that but it
would not have effect in the binary format.