objdump: Add new version of objdump
This is a more advanced version of objdump that what we used to
have. The old version was only dumping data, and while this version
is still a work in progress it mimics very well the behaviour of
GNU objdump, giving very useful debug information. Currently only
coff32 is supported, but elf64 is expected to happen soon.