make: Fix inference updating
The inference() function was used for two different things.
Checking if there was an inference rule that required a
rebuild of a target, and to rebuild a target. Could happen
that a target had to be rebuilt because some other dependency
required it (for example a .o depending of a .h, but the
.c being older than the .o). Adding a parameter to inference()
allows to differentiate between these two cases.