nm: Filter based in know types
Nm by default hides some symbols. We were filtering everything
that is not a function or an object, but we can have symbols
with no type due to a lack of symbolic debug information, and
for that reason is better to filter based in the type of symbols
that we do know that we have to filter.