libmach: Remove Segment data type
The difference between segments and sections
was mainly for elf, which differentiatels between
them, but the attributes from the abstraction point
of view are the same, so we can use only one data
structure but different functions.

As the linker does not merge any sections yet,
we can just comment out the merge function
and nothing changes.