make/posix: Fix feature test macros
_POSIX_C_SOURCE must be defined with a value that signals what
version of POSIX has to be exposed, and just defining it without
any value is equivalent to define it with 0 that is not a valid
version number.

In the same way, posix.c was using S_IFDIR that is XSI, and
to increase the portability it was changed to S_ISDIR that
is defined in the base specification.