libc/darwin: Don't use MB_CUR_MAX for mbstate_t
The definition of mbstate_t was directly copied from the darwin
headers, but it wrongly assumes that MB_CUR_MAX is defined, but
wchar.h does not. The simpliest way to solve the situation is to
hardcode the number 4, because other solutions would need an
internal macro (for example, _MB_CUR_MAX) that is converted to
MB_CUR_MAX if stdlib.h is included. Anyway that solution still
have problems, because this affects to the ABI of darwin, so
our changes to MB_CUR_MAX should not affect to this ABI.