libc/wchar: Rewrite mbtowc()
There are many different corner cases in the implementation of the
underlaying call to mbrtowc(). When mbrtowc() returns -2, it should
be ready to receive new bytes with new calls, and it requires to update
the conversion state.

Also, the standard specifies:

	The implementation shall behave as if no library function calls
	the mbtowc function.

and for that reason, we cannot pass a NULL pointer as state to mbrtowc()
because it would imply to use the hidden state of mbrtowc().

After considering to keep ABI compatibility with the definiton of mbstate_t
in the different systems, it created to many problems. There is 0 guarantees
that code compiled with different libc implementations would work, and for
that reason we dropped the ABI compatibility.