libc/wchar: Group single wchar I/O functions
This makes easier to know the functions that still has to be implemented
and makes easier to know the not implemented functions. It also adds the
macro definition for getwc(). The macro definitions for  getwchar()  and
for putwchar() were considered, but as they required a definition of the
stdin or stdout macros (or the full defition of FILE to have access  to
the __iob array) it was considered worthless and they will just have the
function library definition.