tests/cc: Reuse 0193-incomplete
This test was a copy of 0180-imcomplete, and the issue [i]
stated that using an incomplete array type as parameter of
a function is UB. In our case we accept it, but it makes
more sense to give an error in that case. Rework in 0180
is expected. This test is reused to check that an extern
struct can be declared with incomplete type.

[i]  https://www.open-std.org/jtc1/sc22/wg14/issues/c90/issue0047.html