libc/wchar: Remove state from mbtowc()
As we only support utf8 it does not make sense to keep state in mbtowc()
because as we cannot differentiate between an incorrect sequence and
a non complete sequence then there is not actual use case for that
state. We still keep the static variable because we don't want to use
the internal state from mbrtowc(), and as this function is not safe
thread anyway it is not a problem to have that variable, whose value
will be always correct, except when mbrtowc() returns -2 and in that
case we have to clear it.

The C99 specification does not specify any errno value when mbtowc()
fails, but it allows changes to errno even when it is not directly
specified in the standard, so it is not a problem to implement
mbtowc() using mbrtowc() which sets errno when it cannot translate
a full multibyte sequence.