libc/wchar: Simplify mbsrtowc() and mbstowc()
We only support UTF8, that is stateless, and as mbsrtowc() cannot return
-2 then we do know that the internal state of mbsrtowc() will be always
set to the initial state, so we can reuse it from mbstowcs().

Also, mbsrtowcs() had a bug and it was incrementing the number of wchar
stored in the output buffer when mbrtowc() returned -2. Not a big issue
because we try to read MB_LEN_MAX, and longer sequences will eventually
return -1 and the value of n will not be used.